feat: add cluster license & view kubeconfig
