+++ src/components/CodeEditor/AceEditor.jsx
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2021 KubeClipper Authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+import React from 'react';
+import AceEditor from 'react-ace';
+
+import 'ace-builds/src-noconflict/mode-yaml';
+import 'ace-builds/src-noconflict/mode-groovy';
+import 'ace-builds/src-noconflict/theme-chaos';
+import 'ace-builds/src-noconflict/keybinding-vscode';
+import 'ace-builds/src-noconflict/ext-searchbox';
+
+import './custom.css';
+
+export default function AceEditorWrapper(props) {
+  return (
+    <AceEditor
+      theme="chaos"
+      width="auto"
+      height="100%"
+      tabSize={2}
+      editorProps={{ $blockScrolling: true }}
+      showPrintMargin={false}
+      keyboardHandler="vscode"
+      wrapEnabled
+      {...props}
+    />
+  );
+}
