+++ src/components/CodeEditor/custom.css
@@ -1,0 +1,82 @@
+/*
+ * Copyright 2021 KubeClipper Authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+.ace_editor {
+  line-height: 20px !important;
+  font-family: Monaco, Menlo, Consolas, Courier New, monospace;
+  -webkit-font-smoothing: auto;
+}
+
+.ace_editor.ace-chaos {
+  color: #ffffff;
+  background-color: #242e42;
+}
+
+.ace_editor.ace-chaos .ace_gutter {
+  color: #537f7e;
+  background-color: #242e42;
+  border-right: 1px solid #4a5974;
+}
+
+.ace_editor.ace-chaos .ace_variable,
+.ace_editor.ace-chaos .ace_identifier,
+.ace_editor.ace-chaos .ace_meta.ace_tag {
+  color: #75e0f2;
+}
+
+.ace_editor.ace-chaos .ace_keyword {
+  color: #ffffff;
+}
+
+.ace_editor.ace-chaos .ace_string {
+  color: #ebe087;
+}
+
+.ace_editor.ace-chaos .ace_constant.ace_numeric {
+  color: #bd99ff;
+}
+
+.ace_editor.ace-chaos .ace_marker-layer .ace_active-line {
+  background-color: #36435c;
+}
+
+.ace_editor.ace-chaos .ace_indent-guide {
+  border-right: 1px dotted #777;
+  padding: 2px 0;
+}
+
+.ace_editor.ace-chaos .ace_marker-layer .ace_selection {
+  background-color: #4a5974;
+}
+
+.ace_editor.ace-chaos .ace_comment {
+  color: #aaa;
+}
+
+.ace_editor.ace-chaos .ace_fold:hover {
+  background-color: #fff;
+}
+
+.ace_editor.ace-chaos .ace_line .ace_fold {
+  height: auto;
+}
+
+.ace_search {
+  background-color: #f9fbfd;
+}
+
+.ace_search.right {
+  right: 12px;
+}
