+++ src/components/CodeEditor/index.jsx
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2021 KubeClipper Authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+import React from 'react';
+import PropTypes from 'prop-types';
+import classnames from 'classnames';
+import AceEditor from './AceEditor';
+
+import styles from './index.less';
+
+export default function CodeEditor(props) {
+  const { className, mode, options, value, onChange } = props;
+  return (
+    <AceEditor
+      {...options}
+      className={classnames(styles.editor, className)}
+      value={value}
+      mode={mode}
+      onChange={onChange}
+    />
+  );
+}
+
+CodeEditor.defaultProps = {
+  value: {},
+  mode: 'yaml',
+  options: {},
+  onChange() {},
+};
+
+CodeEditor.propTypes = {
+  value: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.object,
+    PropTypes.arrayOf(PropTypes.string),
+    PropTypes.arrayOf(PropTypes.object),
+  ]),
+  mode: PropTypes.string,
+  options: PropTypes.object,
+  onChange: PropTypes.func,
+};
