+++ src/pages/cluster/components/TerminalCtl/index.less
