+++ src/pages/cluster/containers/Cluster/Detail/Plugins/actions/RemovePlugin.jsx
