+++ src/pages/cluster/containers/Cluster/Detail/Plugins/index.jsx
@@ -19,7 +19,7 @@ import { useRootStore } from 'stores';
 import { get, isEmpty } from 'lodash';
 import { componentStatus } from 'resources/cluster';
 import ActionButton from 'components/Tables/Base/ActionButton';
-import Delete from '../../actions/RemovePlugin';
+import Delete from './actions/RemovePlugin';
 import SaveAsTemplate from '../../actions/SaveAsTemplate';
 import { getAction } from 'utils/allowed';
 import { toJS } from 'mobx';
