+++ src/pages/cluster/containers/Cluster/actions/UpdateLicense.jsx
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2021 KubeClipper Authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+import { ConfirmAction } from 'containers/Action';
+import { rootStore } from 'stores';
+import { checkExpired } from 'utils';
+
+const { clusterStore } = rootStore;
+
+export default class UpdateLicense extends ConfirmAction {
+  get id() {
+    return 'updateLicense';
+  }
+
+  get title() {
+    return t('Update Cluster License');
+  }
+
+  get buttonText() {
+    return t('Update Cluster License');
+  }
+
+  get actionName() {
+    return t('Update Cluster License');
+  }
+
+  policy = 'clusters:edit';
+
+  isLicensExpiration = (item) => checkExpired(item.licenseExpirationTime);
+
+  isRunning = (item) => item.status === 'Running';
+
+  allowedCheckFunc = (item) =>
+    Promise.resolve(this.isRunning(item) && this.isLicensExpiration(item));
+
+  onSubmit = (item) => {
+    const { name } = item;
+    return clusterStore.updateClusterLicense({ name });
+  };
+}
