+++ src/pages/cluster/containers/Cluster/actions/Upgrade.jsx
@@ -19,7 +19,7 @@ import { toJS } from 'mobx';
 import { ModalAction } from 'containers/Action';
 import { rootStore } from 'stores';
 import styles from './index.less';
-import { versionCompare, versionCross } from 'utils';
+import { versionCompare, versionCross, checkExpired } from 'utils';
 import { isIPv4, isDomain, isIpPort } from 'utils/validate';
 
 @observer
@@ -106,7 +106,13 @@ export default class Upgrade extends ModalAction {
     return false;
   }
 
-  static allowed = (item) => Promise.resolve(this.isStatusRunning(item));
+  static isLicensExpiration = (item) =>
+    checkExpired(item.licenseExpirationTime);
+
+  static allowed = (item) =>
+    Promise.resolve(
+      this.isLicensExpiration(item) && this.isStatusRunning(item)
+    );
 
   getMetaVersion(offline) {
     const versionTitle = offline ? 'offlineVersion' : 'onlineVersion';
